added testing route

+++ controllers/blogs.js
@@ -1,9 +1,10 @@
+const jwt = require('jsonwebtoken');
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
 const User = require('../models/user');
 
 blogsRouter.get('/', async (request, response) => {
-	const blogs = await Blog.find({});
+	const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 });
 	response.json(blogs);
 });
 
@@ -21,8 +22,12 @@ blogsRouter.get('/:id', (request, response, next) => {
 
 blogsRouter.post('/', async (request, response, next) => {
 	const body = request.body;
-
-	const user = await User.findById(body.userId);
+	const token = getTokenFrom(request);
+	const decodedToken = jwt.verify(token, process.env.SECRET);
+	if (!token || !decodedToken.id) {
+		return response.status(401).json({ error: 'token missing or invalid' });
+	}
+	const user = await User.findById(decodedToken.id);
 
 	const blog = new Blog({
 		title: body.title,
@@ -40,6 +45,11 @@ blogsRouter.post('/', async (request, response, next) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
+	const token = getTokenFrom(request);
+	const decodedToken = jwt.verify(token, process.env.SECRET);
+	if (!token || !decodedToken.id) {
+		return response.status(401).json({ error: 'token missing or invalid' });
+	}
 	await Blog.findByIdAndRemove(request.params.id);
 	response.status(204).end();
 });
@@ -48,7 +58,11 @@ blogsRouter.put('/:id', async (request, response) => {
 	const body = request.body;
 
 	const blog = {
+		title: body.title,
 		likes: body.likes,
+		url: body.url,
+		user: body.user,
+		author: body.author,
 	};
 
 	const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {
